# Faster los calc signal sum calls for newcalc=False