@pep8speaks 2019-12-02
Hello @Didou09! Thanks for opening this PR. We checked the lines you've touched for [PEP 8](https://www.python.org/dev/peps/pep-0008) issues, and found:





* In the file [`tofu/geom/_core.py`](https://github.com/ToFuProject/tofu/blob/e9af491fb8f046bedbc64fd0a725b8dd09c6fdb4/tofu/geom/_core.py):

> [Line 6183:80](https://github.com/ToFuProject/tofu/blob/e9af491fb8f046bedbc64fd0a725b8dd09c6fdb4/tofu/geom/_core.py#L6183): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (81 > 79 characters)
> [Line 6384:80](https://github.com/ToFuProject/tofu/blob/e9af491fb8f046bedbc64fd0a725b8dd09c6fdb4/tofu/geom/_core.py#L6384): [E501](https://duckduckgo.com/?q=pep8%20E501) line too long (81 > 79 characters)