@lasofivec 2020-01-06
It is not possible to parallelize the code that uses numpy function (GIL-requirement).
So I went for the "hotfix" which is not such a hotfix, more an enhancement.
I tried the new version, which seems to be working, I will create a new PR.