@lasofivec 2020-01-06
Oh, one comment: in your benchmark, you put the parameter `ufunc = [True/False]` which is not a valid parameter for the function. I wonder if you were using the wrong version of tofu ? or a newer that was never pushed ?