@Didou09 2020-01-07
It was a temporary version where all options were possible.
Since the becnhmark showed that using the ufunc (Universal Function, in this case np.add.reduceat) was significantly faster, I went for it and cleaned the function to leave only the ufunc=True implementation for newcalc=False, before pushing. That's why you didn't see it, it was a temporary version for benchmarking before pushing.